+++ packages/react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @augustogomes-ui/react
 
+## 3.0.0
+
+### Major Changes
+
+- add inside button color
+
 ## 2.1.0
 
 ### Minor Changes